Merge rows with the same values
Previously we could only merged a row if none of its non-empty values is
mapped to a non-empty value in the target vector. This commit optimises
this behaviour by allowing the merging of rows if their values are either
mapped to an empty column in the target vector or all non-empty columns
have the same value as in the target vector.