Merge rows in descending order by their number of empty values
This commit adds an optimisation suggested in "Smaller faster table
driven parser" by S. F. Zeigler which proposes to sort rows in a
descending order by their number of empty values. This implementation
doesn't actually sort the input vector, but simply pushes the
displacement value to whatever row was just merged, instead of pushing
the displacement values in a chronological order as before.