Move SparseVec from usize to u64 backing storage.
By using variable sized storage, we couldn't serialise a SparseVec on a
64-bit machine and deserialise on a 32-bit machine (i.e. we couldn't
cross-compile grmtools to WASM). This relatively simple change makes
that possible (since the backing is now always `u64` no matter the
machine).

This is technically an API break (since we've gone from requiring a `T`
compatible with `usize` to a `T` compatible with `u64`), but I doubt
anyone using this library is using a `T` which isn't compatible both
ways.